+++ R/forest.R
@@ -828,7 +828,7 @@
 #'   0.01) or two digits (p > 0.01) (\code{JAMA.pval})
 #' }
 #' Study labels according to JAMA guidelines can be generated using
-#' \code{\link{JAMAlabels}}.
+#' \code{\link{labels.meta}}.
 #' }
 #'
 #' \subsection{Forest plots showing results of subgroups}{
@@ -851,7 +851,7 @@
 #' 
 #' @seealso \code{\link{metabin}}, \code{\link{metacont}},
 #'   \code{\link{metagen}}, \code{\link{forest.metabind}},
-#'   \code{\link{settings.meta}}, \code{\link{JAMAlabels}}
+#'   \code{\link{settings.meta}}, \code{\link{labels.meta}}
 #' 
 #' @references
 #' Guddat C, Grouven U, Bender R, Skipka G (2012):