+++ R/settings.meta.R
@@ -43,7 +43,7 @@
 #' Medical Association}
 #' (\url{https://jamanetwork.com/journals/jama/pages/instructions-for-authors/}).Study
 #' labels according to JAMA guidelines can be generated using
-#' \code{\link{JAMAlabels}}.
+#' \code{\link{labels.meta}}.
 #'
 #' The next commands implement the recommendations of the Institute
 #' for Quality and Efficiency in Health Care, Germany (IQWiG)
@@ -139,7 +139,7 @@
 #' @author Guido Schwarzer \email{sc@@imbi.uni-freiburg.de}
 #' 
 #' @seealso \code{\link{gs}}, \code{\link{forest.meta}},
-#'   \code{\link{print.meta}}, \code{\link{JAMAlabels}}
+#'   \code{\link{print.meta}}, \code{\link{labels.meta}}
 #' 
 #' @examples
 #' # Get listing of current settings